Move back discovery and factsengine to internal
