+++ cmd/config_test.go
@@ -9,8 +9,8 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/stretchr/testify/suite"
 	"github.com/trento-project/agent/internal/agent"
-	"github.com/trento-project/agent/pkg/discovery"
-	"github.com/trento-project/agent/pkg/discovery/collector"
+	"github.com/trento-project/agent/internal/discovery"
+	"github.com/trento-project/agent/internal/discovery/collector"
 )
 
 type AgentCmdTestSuite struct {
