+++ internal/discovery/collector/client.go
