+++ internal/discovery/collector/client_test.go
