+++ internal/discovery/mocks/discovered_cloud_mock.go
