+++ internal/discovery/mocks/discovered_cluster_mock.go
