+++ internal/discovery/mocks/discovered_host_mock.go
