+++ internal/discovery/mocks/discovered_sap_system_mock.go
