+++ internal/discovery/mocks/discovered_subscription_mock.go
