+++ internal/factsengine/adapters/adapter.go
