+++ internal/factsengine/adapters/mocks/Adapter.go
