+++ internal/factsengine/adapters/rabbitmq.go
