+++ internal/factsengine/factsengine_integration_test.go
@@ -9,9 +9,9 @@ import (
 	"golang.org/x/sync/errgroup"
 	"google.golang.org/protobuf/types/known/structpb"
 
-	"github.com/trento-project/agent/pkg/factsengine/adapters"
+	"github.com/trento-project/agent/internal/factsengine/adapters"
+	"github.com/trento-project/agent/internal/factsengine/gatherers"
 	"github.com/trento-project/agent/pkg/factsengine/entities"
-	"github.com/trento-project/agent/pkg/factsengine/gatherers"
 	"github.com/trento-project/contracts/go/pkg/events"
 )
