+++ internal/factsengine/gatherers/corosyncconf.go
