+++ internal/factsengine/gatherers/rcp_plugin.go
