+++ internal/factsengine/gatherers/systemd_test.go
@@ -7,10 +7,10 @@ import (
 	"github.com/coreos/go-systemd/v22/dbus"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/suite"
+	mocks "github.com/trento-project/agent/internal/factsengine/gatherers/mocks"
 	"github.com/trento-project/agent/pkg/factsengine/entities"
-	mocks "github.com/trento-project/agent/pkg/factsengine/gatherers/mocks"
 
-	"github.com/trento-project/agent/pkg/factsengine/gatherers"
+	"github.com/trento-project/agent/internal/factsengine/gatherers"
 )
 
 type SystemDTestSuite struct {
