+++ internal/factsengine/gathering.go
@@ -3,8 +3,8 @@ package factsengine
 import (
 	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
+	"github.com/trento-project/agent/internal/factsengine/gatherers"
 	"github.com/trento-project/agent/pkg/factsengine/entities"
-	"github.com/trento-project/agent/pkg/factsengine/gatherers"
 	"golang.org/x/sync/errgroup"
 )
