+++ internal/factsengine/gathering_test.go
@@ -6,9 +6,9 @@ import (
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/suite"
+	"github.com/trento-project/agent/internal/factsengine/gatherers"
+	"github.com/trento-project/agent/internal/factsengine/gatherers/mocks"
 	"github.com/trento-project/agent/pkg/factsengine/entities"
-	"github.com/trento-project/agent/pkg/factsengine/gatherers"
-	"github.com/trento-project/agent/pkg/factsengine/gatherers/mocks"
 )
 
 type GatheringTestSuite struct {
